feat: Participant 추상화
