+++ src/main/kotlin/blackjack/domain/player/Dealer.kt
@@ -1,34 +1,16 @@
 package blackjack.domain.player
 
-import blackjack.domain.card.PlayingCard
-import blackjack.domain.state.Bust
-import blackjack.domain.state.Ready
-import blackjack.domain.state.Running
-import blackjack.domain.state.State
-
-class Dealer(
-    private var state: State = Ready(),
-) {
-    fun drawCard(card: PlayingCard) {
-        state = state.draw(card)
-    }
-
-    fun cards() = state.hands.cards
-
-    fun stay() {
-        state = state.stay()
-    }
-
-    fun isRunning(): Boolean = state is Running
-
-    fun isBust(): Boolean = state is Bust
-
+class Dealer : Participant() {
     fun needsMoreCard(): Boolean {
         return state.hands.score() <= DRAW_THRESHOLD
     }
 
-    fun score(): Int {
-        return state.hands.score()
+    fun getState(): DealerState {
+        return DealerState(
+            isBust = isBust(),
+            isBlackjack = score() == 21,
+            score = score()
+        )
     }
 
     companion object {
